make silent console private and unique to containerApp
